Many somewhat related changes...
To CTE(): use numerical integration for the Normal Power case; vectorized for confidence levels; arguments harmonized with quantile().

To VaR(): arguments harmonized with quantile().

To exact.R and panjer*.R: use approxfun() instead of stepfun() to build the empirical cdf in order to return 1 past the largest knot.

To simS.R: return the empirical pmf in the environment.

To aggregateDist.R: adjustments to take above change into account; call stored in an attribute of the object instead of inside the environment for better compatibility with ecdf() plot methods.

Various changes to the documentation.

NEWS file updated accordingly.