Add TabBarMode to the distribution
David Holm's TabBarMode, a minor mode to display a tab bar at the top,
similar to the idea of web browser’s tabs is added to the
distribution. Contributed by Rodney Sparapani <rsparapa@mcw.edu>.

Note that TabBarMode is *not* enabled by default.

This changeset was done on macOS to carry over the changes from
emacs-modified-macos. It is not tested on Windows.